CXP-1182: Move and rename command queries validation test
