+++ components/catalogs/back/tests/Integration/ServiceAPI/Command/CreateCatalogCommandValidationTest.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Akeneo\Catalogs\Test\Integration\Application\Handler;
+namespace Akeneo\Catalogs\Test\Integration\ServiceAPI\Command;
 
 use Akeneo\Catalogs\ServiceAPI\Command\CreateCatalogCommand;
 use Akeneo\Catalogs\Test\Integration\IntegrationTestCase;
@@ -12,7 +12,7 @@ use Symfony\Component\Validator\Validator\ValidatorInterface;
  * @copyright 2022 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  */
-class CreateCatalogHanderTest extends IntegrationTestCase
+class CreateCatalogCommandValidationTest extends IntegrationTestCase
 {
     private ?ValidatorInterface $validator;
